feat: Category用のグローバルキャッシュの実装
